Check for Mac::Finder::DSStore in scripts/osx-dmg.sh and spit out a warning + url if it does not exist.
Originally committed to SVN as r2584.

+++ scripts/osx-dmg.sh
@@ -21,6 +21,13 @@ if ! test -d "${PKG_DIR}"; then
   exit 1;
 fi
 
+if ! perl -e 'require Mac::Finder::DSStore' > /dev/null 2>&1; then
+  echo
+  echo "Perl Mac::Finder::DSStore is required to build a dmg."
+  echo "Please get it from http://freehg.org/u/wiml/dsstore/"
+  exit;
+fi
+
 rm -rf ${TMP_DMG} "${PKG_NAME}.dmg"
 mkdir -v ${TMP_DMG}
 echo
